Split stage data up to yaml files (Fix #112)
