+++ api/seeders/20161003085039-first-seed.js
@@ -1,4 +1,4 @@
-const stages = require('../../lib/stages.json');
+const stages = require('../../stages');
 const Stages = require('../models/stage');
 
 module.exports = {
