+++ lib/index.js
@@ -5,7 +5,7 @@ require('core-js/es6');
 
 const querystring = require('querystring');
 const $ = require('jquery');
-const stageConfigs = require('./stages.json');
+const stageConfigs = require('../stages');
 const Game = require('./game');
 
 // parse GET parameters of the location
