+++ lib/util.js
@@ -1,4 +1,5 @@
 const typeToConfig = require('./type-to-config');
+// const stages = require('../stages');
 const blockNames = Object.keys(typeToConfig);
 
 // Check the equality of two arrays *as sets*
@@ -44,9 +45,11 @@ module.exports.validateSubmission = (submission) => {
 		return {pass: false, message: 'board should be an array'};
 	}
 
-	let error = null;
+	if (submission.board.length === 0) {
+		return {pass: false, message: 'board must at least have one block'};
+	}
 
-	// TODO: should at least have one block
+	let error = null;
 
 	submission.board.forEach((block, index) => {
 		const message = (() => {
