+++ stages/100-again.yml
@@ -1,0 +1,24 @@
+name: 100-again
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  const-1: null
+  add: null
+  sub: null
+  mul: null
+inputX: 2
+outputX: 2
+input:
+  - -1
+  - 10
+  - 1
+output:
+  - 100
+  - 100
+  - 100
+width: 5
+height: 5
+statement: 100を作ろう!
+title: 100再び
