+++ stages/100.yml
@@ -1,0 +1,21 @@
+name: '100'
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  equal: null
+  add: null
+inputX: 4
+outputX: 4
+input:
+  - 10
+  - 0
+  - -1
+output:
+  - 100
+  - 100
+  - 100
+width: 9
+height: 8
+statement: 100を作ろう!
+title: '100'
