+++ stages/calc01.yml
@@ -1,0 +1,25 @@
+name: calc01
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-2: null
+inputX: 1
+outputX: 1
+input:
+  - 8
+  - 3
+  - 9
+output:
+  - 16
+  - 6
+  - 18
+width: 3
+height: 3
+statement: 数を2倍してみよう!
+title: 四則演算 -基本- 1
+modal: '02'
