+++ stages/calc02.yml
@@ -1,0 +1,21 @@
+name: calc02
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  plus-1: null
+  minus-2: null
+inputX: 2
+outputX: 2
+input:
+  - 8
+  - 15
+  - 10
+output:
+  - 5
+  - 12
+  - 7
+width: 5
+height: 5
+statement: 数を3引いてみよう!
+title: 四則演算 -基本- 2
