+++ stages/calc03.yml
@@ -1,0 +1,24 @@
+name: calc03
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-2: null
+inputX: 2
+outputX: 2
+input:
+  - 1
+  - 3
+  - 8
+output:
+  - 6
+  - 18
+  - 48
+width: 5
+height: 5
+statement: 数を6倍してみよう!
+title: 四則演算 -基本- 3
