+++ stages/calc04.yml
@@ -1,0 +1,24 @@
+name: calc04
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-2: null
+inputX: 2
+outputX: 2
+input:
+  - 8
+  - 3
+  - 9
+output:
+  - 19
+  - 9
+  - 21
+width: 5
+height: 5
+statement: 数を2倍して、3足してみよう!
+title: 四則演算 -基本- 4
