+++ stages/calc05.yml
@@ -1,0 +1,21 @@
+name: calc05
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: 1
+  plus-1: 1
+inputX: 2
+outputX: 2
+input:
+  - 8
+  - 5
+  - 3
+output:
+  - 18
+  - 12
+  - 8
+width: 5
+height: 5
+statement: 数を2倍して、2足してみよう!
+title: 四則演算 -基本- 5
