+++ stages/calc06.yml
@@ -1,0 +1,20 @@
+name: calc06
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  add: null
+inputX: 1
+outputX: 1
+input:
+  - 8
+  - 3
+  - 9
+output:
+  - 16
+  - 6
+  - 18
+width: 3
+height: 2
+statement: 数を2倍してみよう!
+title: 四則演算 -基本- 6
