+++ stages/complement-of-2.yml
@@ -1,0 +1,52 @@
+name: complement-of-2
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  wireX: null
+  times-2: null
+  times-3: null
+  div-2: null
+  div-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  const-0: null
+  const-1: null
+  const-2: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  mod: null
+  pow: null
+  bitshift-left: null
+  bitshift-right: null
+  bitwise-and: null
+  bitwise-or: null
+  bitwise-xor: null
+  equal: null
+  neq: null
+  gt: null
+  geqq: null
+  lt: null
+  leqq: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 3
+outputX: 3
+input:
+  - 1
+  - 64
+  - 127
+output:
+  - 255
+  - 192
+  - 129
+width: 7
+height: 7
+statement: 8bitの2の補数を計算してみよう!
+title: 2の補数
