+++ stages/conditional02.yml
@@ -1,0 +1,32 @@
+name: conditional02
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-2: null
+  equal: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  c-contact: null
+  conditional: null
+  transistor: null
+inputX: 5
+outputX: 5
+input:
+  - 5
+  - 6
+  - 7
+output:
+  - 10
+  - 18
+  - 7
+width: 11
+height: 11
+statement: ifを使って、5は2倍、6は3倍、それ以外は1倍をしてみよう!
+title: 条件分岐 -基本- 2
