+++ stages/division-easy.yml
@@ -1,0 +1,29 @@
+name: division-easy
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+inputX: 2
+outputX: 2
+input:
+  - 8
+  - 18
+  - 6
+output:
+  - 4
+  - 9
+  - 3
+width: 5
+height: 5
+statement: ➗ブロックをうまく使って、数を2で割ってみよう!
+title: 割り算 -easy-
