+++ stages/division-hard.yml
@@ -1,0 +1,28 @@
+name: division-hard
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+inputX:
+  - 1
+  - 3
+outputX: 2
+input:
+  - - 3
+    - 6
+  - - 71
+    - 7597
+  - - -13
+    - 117
+output:
+  - 2
+  - 107
+  - -9
+width: 5
+height: 5
+statement: 右の数を左の数で割ってみよう!
+title: 割り算 -hard-
