+++ stages/factorial.yml
@@ -1,0 +1,33 @@
+name: factorial
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-2: null
+  equal: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 6
+outputX: 6
+input:
+  - 3
+  - 5
+  - 7
+output:
+  - 6
+  - 120
+  - 5040
+width: 13
+height: 13
+statement: 階乗を計算してみよう!
+title: 階乗
