+++ stages/fibonacci.yml
@@ -1,0 +1,34 @@
+name: fibonacci
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  times-2: null
+  times-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  equal: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 5
+outputX: 5
+input:
+  - 5
+  - 8
+  - 13
+output:
+  - 5
+  - 21
+  - 233
+width: 11
+height: 11
+statement: n番目のフィボナッチ数を計算してみよう!
+title: フィボナッチ数
