+++ stages/gcd.yml
@@ -1,0 +1,51 @@
+name: gcd
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  wireX: null
+  times-2: null
+  times-3: null
+  div-2: null
+  div-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  const-0: null
+  const-1: null
+  const-2: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  mod: null
+  equal: null
+  neq: null
+  gt: null
+  geqq: null
+  lt: null
+  leqq: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX:
+  - 5
+  - 7
+outputX: 6
+input:
+  - - 6
+    - 8
+  - - 30
+    - 12
+  - - 432
+    - 126
+output:
+  - 2
+  - 6
+  - 18
+width: 13
+height: 13
+statement: 二数の最大公約数を計算してみよう!
+title: 最大公約数
