+++ stages/index.js
@@ -1,0 +1,43 @@
+/* eslint global-require: 'off' */
+
+module.exports = [
+	require('./wire01.yml'),
+	require('./calc01.yml'),
+	require('./calc02.yml'),
+	require('./calc03.yml'),
+	require('./calc04.yml'),
+	require('./calc05.yml'),
+	require('./calc06.yml'),
+	require('./make-minus-one-easy.yml'),
+	require('./division-easy.yml'),
+	require('./sixth-power.yml'),
+	require('./remainder.yml'),
+	require('./plus-32.yml'),
+	require('./make-minus-one-med.yml'),
+	require('./conditional01.yml'),
+	require('./conditional02.yml'),
+	require('./conditional03.yml'),
+	require('./factorial.yml'),
+	require('./parity.yml'),
+	require('./fibonacci.yml'),
+	require('./bivariation01.yml'),
+	require('./power-easy.yml'),
+	require('./power-hard.yml'),
+	require('./division-hard.yml'),
+	require('./gcd.yml'),
+	require('./lcm.yml'),
+	require('./make-minus-one-hard.yml'),
+	require('./sqrt-easy.yml'),
+	require('./sqrt-hard.yml'),
+	require('./complement-of-2.yml'),
+	require('./binarian-easy.yml'),
+	require('./perfect-number.yml'),
+	require('./reversal.yml'),
+	require('./msd.yml'),
+	require('./mod3-hard.yml'),
+	require('./max.yml'),
+	require('./the-fifth-max.yml'),
+	require('./100.yml'),
+	require('./100-again.yml'),
+	require('./1000.yml'),
+];
