+++ stages/mod3-hard.yml
@@ -1,0 +1,44 @@
+name: mod3-hard
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  wireX: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  times-2: null
+  times-3: null
+  const-0: null
+  const-1: null
+  const-2: null
+  add: null
+  sub: null
+  pow: null
+  equal: null
+  neq: null
+  gt: null
+  geqq: null
+  lt: null
+  leqq: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 9
+outputX: 9
+input:
+  - 12345678
+  - 31415926
+  - 27182818
+  - 77777777
+output:
+  - 0
+  - 1
+  - 1
+  - 2
+width: 19
+height: 19
+statement: 3で割った余りを求めてみよう!
+title: mod3 -hard-
