+++ stages/perfect-number.yml
@@ -1,0 +1,52 @@
+name: perfect-number
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  wireX: null
+  times-2: null
+  times-3: null
+  div-2: null
+  div-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  const-0: null
+  const-1: null
+  const-2: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  mod: null
+  pow: null
+  bitshift-left: null
+  bitshift-right: null
+  bitwise-and: null
+  bitwise-or: null
+  bitwise-xor: null
+  equal: null
+  neq: null
+  gt: null
+  geqq: null
+  lt: null
+  leqq: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 2
+outputX: 2
+input:
+  - 3
+  - 31
+  - 8191
+output:
+  - 6
+  - 496
+  - 33550336
+width: 5
+height: 5
+statement: メルセンヌ素数と完全数の関係を考えてみよう!
+title: 完全数
