+++ stages/sqrt-easy.yml
@@ -1,0 +1,43 @@
+name: sqrt-easy
+parts:
+  wireI: null
+  wireL: null
+  wireT: null
+  wireX: null
+  times-2: null
+  times-3: null
+  div-2: null
+  div-3: null
+  plus-1: null
+  plus-2: null
+  minus-1: null
+  minus-2: null
+  const-0: null
+  const-1: null
+  const-2: null
+  add: null
+  sub: null
+  div: null
+  mul: null
+  mod: null
+  pow: null
+  equal: null
+  neq: null
+  c-contact: null
+  conditional: null
+  transistor: null
+  diode: null
+inputX: 4
+outputX: 4
+input:
+  - 9
+  - 16
+  - 36
+output:
+  - 3
+  - 4
+  - 6
+width: 9
+height: 9
+statement: 平方根を計算してみよう!
+title: 平方根 -easy-
