+++ stages/wire01.yml
@@ -1,0 +1,18 @@
+name: wire01
+parts:
+  wireI: null
+inputX: 1
+outputX: 1
+input:
+  - 4
+  - 10
+  - 7
+output:
+  - 4
+  - 10
+  - 7
+width: 3
+height: 3
+statement: 入り口と出口をつなげてみよう!
+title: 練習
+modal: '01'
